requirements.txt
# RAG-Food System Dependencies
# Updated: October 16, 2025

# Vector Database - Upstash (Cloud)
upstash-vector>=0.4.0

# Configuration Management
python-dotenv>=1.0.0

# HTTP Requests
requests>=2.31.0

# Optional: Groq Cloud API (for LLM generation)
groq>=0.4.0

# Standard Library (included with Python)
# - json
# - os

# Development Dependencies (optional)
# pytest>=7.0.0
# pytest-cov>=4.0.0
